bitcore-hardware utility class
