+++ packages/bitcore-hardware/src/types/base.ts
